UniPMTService/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
fcd-torch==1.0.7
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
h5py==2.10.0
huggingface-hub==0.30.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
molsets==0.1.3
multidict==6.1.0
networkx==2.6.2
numpy==1.20.3
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pip==24.2
propcache==0.2.0
psutil==7.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
scikit-learn==0.24.2
scipy==1.6.2
seaborn==0.13.2
setuptools==75.1.0
six==1.17.0
threadpoolctl==3.5.0
tokenizers==0.13.3
torch_cluster==1.6.3
torch-geometric==2.6.1
torch_scatter==2.1.2
torch_sparse==0.6.18
torch-spline-conv==1.2.2
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.67.1
transformers==4.29.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wheel==0.44.0
yarl==1.15.2
zipp==3.20.2
